moved speedup from experiments
